Premium Analytics: align Top posts widget structure with block conventions
Address PR #49568 review feedback:

* Collapse the widget into render.tsx (matching the canonical hello-world
  reference). Remove top-posts-widget.tsx, index.ts and types.ts; rename the
  presentational TopPostsWidget component to TopPostsLeaderboard and keep it
  internal. A widget is the block-level abstraction (widget.json + widget.ts +
  render.tsx), so render.tsx is the component itself, like a block's edit.js.
* Stop exporting TopPostsAttributes / TopPostRow (used only here).
* Rename top-posts-widget.module.css -> top-posts.module.css.
* i18n: wrap widget.ts title and attribute labels in __().
* Replace the native anchor with the @wordpress/ui Link component
  (openInNewTab + unstyled variant to preserve the leaderboard row styling).
* Defer the Storybook story to the #49505 follow-up; drop its stale changelog.
