--- projects/packages/premium-analytics/changelog/add-top-posts-widget-storybook
@@ -1,4 +1,0 @@
-Significance: patch
-Type: added
-
-Storybook: Add a prop-driven Top posts & pages widget story under Packages/Premium Analytics/Widgets.
