--- projects/packages/premium-analytics/widgets/top-posts/index.ts
@@ -1,2 +1,0 @@
-export { TopPostsWidget, type TopPostsWidgetProps } from './top-posts-widget';
-export type { TopPostRow } from './types';
